+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Adds support for `setOnScrollPositionChange` method to the `AndroidWebViewController`.
 
+## 3.12.1
+
+* Fixes `use_build_context_synchronously` lint violations in the example app.
+
 ## 3.12.0
 
 * Adds support for `PlatformWebViewController.getUserAgent`.